renewal refactor
