+++ src/pages/renew/hooks/useRenewHandlers.ts
@@ -1,0 +1,10 @@
+import { useState } from "react";
+
+export const useRenewHandlers = () => {
+  const [activeIdx, setActiveIdx] = useState<number>(0);
+
+  return {
+    activeIdx,
+    setActiveIdx,
+  }
+}
